test fix build

+++ packages/widget/src/utils/misc.tsx
@@ -23,6 +23,7 @@ export const copyToClipboard = (string?: string) => {
 export function atomWithStorageNoCrossTabSync<T>(storageKey: string, initialValue: T) {
   const defaultStorage: SyncStorage<T> = {
     getItem: (key) => {
+      if (!localStorage) return;
       const storedValue = localStorage.getItem(key);
       if (!storedValue) return initialValue;
 
@@ -33,9 +34,11 @@ export function atomWithStorageNoCrossTabSync<T>(storageKey: string, initialValu
       }
     },
     setItem: (key, newValue) => {
+      if (!localStorage) return;
       localStorage.setItem(key, JSON.stringify(newValue));
     },
     removeItem: (key) => {
+      if (!localStorage) return;
       localStorage.delete(key);
     },
   };
